Add files via upload
In multiple related datasets, if you need to subset dataset 2 retaining all rows whose indices are the same as those in dataset 1 where a condition is satisfied and apply a function to them, you could use this. This should be expanded such that instead of row indices, we could do a join based on a foreign key for all rows in dataset 1 that satisfy a condition (in this case match a search string).